ml-worker/extract-line-chart-data/requirements.txt
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
pandas==2.1.4
Pillow==10.2.0
opencv-python-headless==4.9.0.80
transformers==4.36.2
huggingface-hub==0.20.2
openmim==0.3.9
terminaltables==3.1.10
python-dotenv==1.0.0
psycopg2-binary==2.9.9
paho-mqtt==2.1.0
requests==2.31.0
PyYAML==6.0.1
tqdm==4.66.1
scikit-image==0.22.0
bresenham==0.2
git+https://github.com/tdsone/LineFormer.git@ea059eae00285657e5668217bdd48e5435a9800e
